requirements.txt
asgiref==3.3.1
Django==3.1.5
django-cors-headers==3.6.0
django-money==1.2.2
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
phonenumbers==8.12.15
py-moneyed==0.8.0
python-decouple==3.4
pytz==2020.5
sqlparse==0.4.1
